+++ src/pages/ResidentPage.js
@@ -5,8 +5,8 @@ import Tooltip from 'react-bootstrap/Tooltip';
 import ResidentGrid from '../components/Grids/ResidentGrid';
 import ResidentEdit from '../components/Modals/ResidentEdit';
 import {FullName} from '../utility/common';
-import RefreshMedicineList from "../providers/RefreshMedicineList";
-import RefreshMedicineLog from "../providers/RefreshMedicineLog";
+import RefreshMedicineList from "../providers/helpers/RefreshMedicineList";
+import RefreshMedicineLog from "../providers/helpers/RefreshMedicineLog";
 import ConfirmationDialog from "../components/Modals/ConfirmationDialog";
 import {Form} from "react-bootstrap";
 import PropTypes from 'prop-types';
